# docs/swagger.yaml
basePath: /v0
definitions:
...
  inputs.AlertGroupAlerts:
    properties:
      alert_ids:
        items:
          type: integer
        type: array
    required:
    - alert_ids
    type: object
  inputs.AlertGroupBody:
    properties:
      alert_ids:
        items:
          type: integer
        type: array
      is_dafault:
        default: false
        type: boolean
      name:
        type: string
    required:
    - alert_ids
    - name
    type: object
  inputs.AlertGroupUpdate:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  models.Response:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: I have no specific description
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hod Hod API
  version: "1.0"
paths:
...
  /alert-group:
    get:
      description: return AlertGroups
      operationId: get_user_alertGroups
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get AlertGroups
      tags:
      - Alert Group
    post:
      description: new AlertGroup
      operationId: new_AlertGroup
      parameters:
      - description: alert_group
        in: body
        name: alert-group
        required: true
        schema:
          $ref: '#/definitions/inputs.AlertGroupBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: new AlertGroup
      tags:
      - Alert Group
  /alert-group/{id}:
    delete:
      description: delete alertGroup by alertGroupID
      operationId: delete_alert_group
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: delete alertGroup
      tags:
      - Alert Group
    get:
      description: return AlertGroup
      operationId: get_alert_Group
      parameters:
      - description: alert_grp_id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get AlertGroup
      tags:
      - Alert Group
    put:
      description: update AlertGroup
      operationId: update_AlertGroup
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: alert_group
        in: body
        name: alert_group
        required: true
        schema:
          $ref: '#/definitions/inputs.AlertGroupUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: update AlertGroup
      tags:
      - Alert Group
  /alert-group/{id}/alert:
    put:
      description: update AlertGroupAlerts
      operationId: update_AlertGroupAlerts
      parameters:
      - description: alert_group_id
        in: path
        name: id
        required: true
        type: integer
      - description: alerts
        in: body
        name: alerts
        required: true
        schema:
          $ref: '#/definitions/inputs.AlertGroupAlerts'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: update AlertGroupAlerts
      tags:
      - Alert Group
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"